refactor: move PublicFormProvider into own component
it's going to start becoming big, so we should make it more searchable
